Add files via upload
These files were used for the initial project. They may have been adopted for the later publication and the plots were changed for a comprehensive illustration of the data.

The given files can help to create data with an installed OpenLCA tool, update data (in csv format) and create a comparison between different vehicles and different drive train types. The results can be read in the following article:
https://www.mdpi.com/2032-6653/15/3/114